[activation] Merge ActivationType.py in Activation.py
