+++ Tensile/ClientWriter.py
@@ -25,7 +25,6 @@ from . import LibraryIO
 from .Common import globalParameters, pushWorkingPath, popWorkingPath, print1, printExit, CHeader, printWarning, listToInitializer, ClientExecutionLock
 from .SolutionStructs import ProblemType, ProblemSizesMock
 from .TensileCreateLibrary import copyStaticFiles
-from .ActivationType import ActivationType
 
 import os
 import subprocess
