+++ Tensile/KernelWriterActivationEnumHeader.py
@@ -22,7 +22,7 @@
 from copy import deepcopy
 
 from .Common import globalParameters, CHeader
-from .ActivationType import ActivationType
+from .Activation import ActivationType
 from .KernelWriterBase import KernelWriterBase
 
 class KernelWriterActivationEnumHeader(KernelWriterBase):
